backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
bcrypt
PyJWT
pydantic
email-validator
psycopg2-binary
